+++ src/cadso-ui-component-test-example/__tests__/test.cadso-ui-component-test-example.js
@@ -1,4 +1,4 @@
-describe('cadso-ui-component-test-example Test', () => {
+describe('Simple Working Test', () => {
 	it('should be true', () => {
 		expect(true).toBe(true);
 	});
